+++ sdk/resources/azure-mgmt-resource/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History
 
+## 10.3.0 (2020-11-02)
+**Features**
+
+  - Model DeploymentWhatIfProperties has a new parameter expression_evaluation_options
+  - Model DeploymentProperties has a new parameter expression_evaluation_options
+  - Model TemplateSpec has a new parameter versions
+  - Added operation ProvidersOperations.register_at_management_group_scope
+  - Added operation ApplicationDefinitionsOperations.get_by_id
+  - Added operation ApplicationDefinitionsOperations.create_or_update_by_id
+  - Added operation ApplicationDefinitionsOperations.delete_by_id
+  - Added operation group ApplicationClientOperationsMixin
+
 ## 10.2.0 (2020-08-03)
 
 **Features**